local/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-documentai
pandas
PyMuPDF
requests
python-dotenv
openpyxl